requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.30.0
cozepy>=0.6.2
